fix: resolve lint errors
